Don't re-raise JSON::ParserError in Snapshotter service
This was originally a bit of defensive programming to hopefully help
surface problems that would happen during the running of the
take-snapshot.js script. However, when run on a URL that had JavaScript
errors, this would prevent the snapshot from being taken. This isn't a
great thing, so instead of re-raising these things, I am changing it so
they are simply swallowed.

We may want to collect these lines and output them on the snapshot page,
but for the time being that is YAGNI.

Change-Id: Id0068e4626e00bf8a209a299e9d8aa319da02226