Add j, k, o, and Enter keyboard shortcuts
We want Diffux to be easy for power users to use. For us, this means
convenient keyboard shortcuts are a must. To that end, this commit sets
up an easy to deploy mechanism for using j, k, o, and Enter keyboard
shortcuts to navigate between focusable items on the page.

In short, j and k will focus the next or previous focusable element. o
or Enter will click it if it is a link, or will click the first
descendant link within the focused element. To specify which elements
are focusable, simply add the data-keyboard-focusable attribute to the
element. The currently focused element is given a .keyboard-focused
class.

This commit deploys this to the project listing page and only includes
styling for focused table rows, but I will soon follow up with some
other places and clean out my previous attempt at keyboard shortcuts.

This is in service of #13.

Change-Id: I4103c7f27faf5098420782d6872bfe173eb9482c