Replace snapshot keyboard navigation with more general solution
In 655263bf12837ef I added code that allows us to use
data-keyboard-focusable attributes to specify which elements on a page
are focusable via keyboard shortcuts j and k, and openable via keyboard
shortcuts o and Enter. This commit completes some of the cleanup
promised there by partially replacing the system we were using for
navigation between snapshots within a sweep with this new, more general
system. To finish this work, I'll be adding keyboard shortcuts for [ and
] which perform a j or k, respectively, immediately followed by an o
keyboard shortcut.

Change-Id: Ia7db2e2beedaef6773c032d46c2c09f948ca263c