Scroll to focused element when using keyboard shortcuts
When using the keyboard to navigate between focusable elements, the
focusable element you end up focusing may end up being outside of the
current viewport. This makes this feature difficult to use. Therefore,
after focusing a focusable element, we should always tell the browser to
make sure that element is scrolled within view.

This is in service of #13.

Change-Id: I4cd65603dd08028703e94d7bc426f37bb02be97c