Revert side-scroll for "In review" panel
The "In review" panel on a snapshot page used to have horizontal
scrolling to save space. It was hard to make keyboard navigation to work
with this as `scrollIntoFocus` didn't make the currently highlighted
item selected.

In service of keeping things lean and simple, I'm reverting this
decision. This means reverting commits 2166537, 452802d, and
595e547.