Improve scrolling on held-down j/k key
The jQuery `stop()` method takes two parameters that we can use to make
the scrolling experience better when you are keyboard-navigating and
holding down j or k. The first boolean argument cancels all queued
animations. That's probably not too important for us. The second param
however, controls if the animation should be fast-forwarded when
cancelled (jump to the end). With that set to `true`, scrolling is way
faster when you hold down j or k.