Report diff clusters when comparing snapshots
This commit introduces the DiffClusterFinder service, a class
responsible for finding clusters of differences. A cluster is defined as
adjacent rows of pixels that have at least one pixel different. A row is
adjacent to another row if it is closer than MAXIMUM_ADJACENCY_GAP
pixels (currently 20).

An upcoming commit will persist this information, and use it when
navigating the snapshot diff to highlight clusters.