Persist diff clusters
The previous commit, 2fb0a5d, introduced a way to find clusters in a
diff. This commit makes use of that information and persists each
cluster to a SnapshotDiffCluster model.

Apart from storing the clusters, I also added an `image_height` field to
the SnapshotDiff model. This will come in handy when in an upcoming
commit when I will make use of this information to highlight clusters in
the diff.