Make diff clusters keyboard-focusable
Now that we have information about where all the diff clusters are, we
can use it to enhance the snapshot review process. I've added each
cluster to the diff image container and made them keyboard-focusable.
This means that you can quickly scan through a diff using j and k. The
currently focused cluster will show up as a 2 pixel bright border to the
right border of the diff image.