Move some ruby code out of snapshot `#show` template
I added some percentage calculation to the
app/views/snapshots/show.html.haml template in a recent commit. That
code belongs in a ruby file so I moved it to the model. The model is
still thin, so even if the new methods could be considered helpers I
think it's fine to keep them here for a while.