Change keyboard shortcuts for `[` and `]`
These keys used to move up/down once and then auto-click the currently
focused link. Now that we've moved to pager ("Previous"/"Next") buttons
instead of the "In review" panel, these shortcuts didn't really do much.

For now, I've changed to a specific solution (looks for links in a
`.pager` container). In the future we might consider finding a more
general solution to these shortcut keys.