Refactor breadcrumbs into partial
We had a lot of duplicated markup related to breadcrumbs so I refactored
that code to render through a partial instead. This sets us up for some
improvements to keyboard shortcuts, I have a goal of adding something
like `data-keyboard-shortcut="u"` to create a keyboard shortcut on the
`u` key for that particular target. That's a future commit though.