Fix keyboard shortcuts for "[" and "]"
In 4c043d25c1fed7ab183e we generalized keyboard shortcuts to add data
attributes to elements for keyboard shortcuts. This would then do a
data-attribute query against the DOM to find the data-attribute that
contained the pressed key. However, since attribute queries use square
brackets, the way the query was written ended up being invalid when the
keyboard shortcut also used square brackets, causing the following error
to appear in the console:

  Uncaught Error: Syntax error, unrecognized expression:
  [data-keyboard-shortcut=]]

To fix this, I surrounded the value for the selector in double quotes.
This solution obviously has the same problem if we choose to use a
double quote for a keyboard shortcut, but for the time being any more
advanced escaping is YAGNI.

Change-Id: I9d1f05bf6b6bc7bc88a4dbd4ca0d9d22245177c4