Make DB connection pool as large as Sidekiq worker pool
I was reading through the Sidekiq documentation and found that they
recommend that the number of connections in the database pool closely
match the number of Sidekiq workers run (which by default is 25). This
commits makes that happen.

After applying this commit, you should see fewer errors in the Sidekiq
console log about not being able to open a database connection.