Change "changed" color from magenta to violet
The magenta we were using was visually too close to the red we were
using. When these colors were next to each other, it was difficult to
tell them apart. This commit attempts to make it easier to see the
difference between these two colors by changing magenta to violet. I had
considered changing red, but since it has significant meaning baked into
it, it seemed that magenta was the best choice to change.

Change-Id: I21900de77530e9afc4e1f67aaf65b931424a34d4