Fix swapped before/after images
When I first implemented this, I had a strange issue where rows that
were `#added?` were actually deleted, and rows that were `#deleted?`
were actually added. At the time I did the quick fix of just swapping
the colors, since we treat these the same except for this one color.
This commit fixes it further upstream by swapping the before and after
images that are sent to sdiff.

Change-Id: I793b02928be833f77431f04c9a40a474737ab1c1