Simplify projects#show page
@lencioni and I started looking into speeding up the project#show page.
We were doing a lot of N+1 queries and in dev it was taking up to a
second to render this page on the server.

After going down the path of preloading to reduce N+1 we decided that it
probably wasn't worth the effort, and instead started talking about how
to make the page simpler, but still support the main use cases. We
decided to go with rendering the list of Urls in a table and not show
any snapshot cards at all.

This reverts part of 2633988f.