Introduce stacked progress bar for sweeps
To be able to quickly see the status for a sweep, I'm introducing a
progress bar that displays the progress. The progress bar is stacked in
the following order:

1. Under review
2. Accepted
3. Rejected

I choose to keep the old `sweep_status` method and use it to display
a readable status format when you hover over a progress bar.

The progress bar is displayed in the following places:

- On the project page, under "Last sweep"
- On the sweeps#index page, in the table
- On the sweeps#show page