Simplify sweeps#show page
Now that we have a progress bar to display the status for a sweep, we
don't really need to split snapshots into groups on the sweeps#show
page. This commit simplifies rendering by putting all snapshots in one
panel.

This breaks a previous feature. Before, only snapshots in the "Under
review" pane would take you to the snapshots#show page in the context of
reviewing a list of snapshots. After this commit, all snapshots will
take you to the snapshots#show page in a review context. Also, all
snapshots in the sweep will be part of the review list. We believe this
to be a better experience as it is less state-dependent. Independent of
what state individual snapshots are in, the review order and size will
remain.

This change is also in preparation for adding some real-time updating
features to the page. I'm about to make snapshot cards and sweep
progress bars auto-update, and I don't want content to have to shift
around between different containers (which would have been the case
before this change).