Auto-refresh snapshot cards every 5 seconds
This commits adds a 5 second polling component to automatically refresh
snapshot cards. The basic setup:

- The browser will poll/refresh every five seconds, passing ids for
  all snapshot cards that can be found on the page.
- The server will lookup snapshots by ids and render a new snapshot
  card. The new card is sent back in a json array.
- The browser parses this json and replaces cards with new ones.

"What about web sockets?"

I'm fully aware that this type of problem could be solved by using
websockets. I've built similar components before [1] and even though it
might seem nice at the surface, websockets introduce complexities that
I'd much rather be without at this time. First of all, the server needs
to keep track of clients. Then, to avoid pushing out every single change
to everyone, we would need to keep track of what the clients are
currently showing on the screen and only push out relevant push messages
to those who need them.

With client-side polling, the interface is simple: The client constructs
a request with the elements it wants to refresh. The server responds
with new versions of those elements. No state needs to be kept on the
server.

This component is set up to be able to handle multiple types (snapshots,
sweeps) at the same time in one request. Right now we only do snapshots,
but sweep progress bars are coming up.

To avoid re-rendering all cards every time we poll, I chose to pass
through an `ifModifiedSince` timestamp. This timestamp is set globally
on page load, and updated each time we poll. Only snapshots changed
after this timestamp will be re-rendered.

[1]
https://github.com/trotzig/Burrito/blob/8eba6e60f67b6bd059df1a6982049feafcf0f0f9/src/main/resources/burrito/public/burrito-feeds.js