Auto-refresh sweep progress bars
Now that we have a general refresh-pattern for components, I've added
sweep progress bars to the list of items that are automatically
refreshed.

I generalized the controller code a bit to detach the different types
(snapshots, sweeps) from the action itself. This was needed in order to
keep the code DRY, and as an added benefit it should now be easier to
add other elements to refresh.

Specs are at this point mostly smoke tests. We might want to test the
actual html output from the refresh. But for now I've left that out of
scope.