Make snapshot cards have fixed width
Now that we are auto-refreshing snapshot cards, we don't want them to
move around on the screen. Before, cards could be anything from 100
to 120 pixels. After this change, they are always 110 pixels.

I also changed the text color of the timestamp and the diff percentage,
plus I rounded the percentage down to the closes two decimal value.