Always show snapshot in context of sweep
Before, we used to only show the next/previous buttons on the snapshot
page if you hit a link to the page from the sweeps#show page. Now that
we group all snapshots for a sweep in the same panel and show all
snapshots as part of the review-list, we no longer need the complexity
of having to pass along the snapshot ids of the snapshots currently in
review. Instead, we can construct the review-list from all snapshots for
the sweep that the currently shown snapshot belongs to.

This fixes a bug where an auto-refreshed snapshot on a sweep page would
have a link to a review-list-less snapshot page.