Make SnapshotComparer combine three images into one
This commit breaks the SnapshotComparer apart and delegates the image
creation to a new model, SnapshotComparisonImage, and its three
subtypes:

    SnapshotComparisonImage::Before - knows how to render the
    before-image version of the snapshot.

    SnapshotComparisonImage::Overlayed - renders a combined image
    from the before and after snapshot.

    SnapshotComparisonImage::After - creates the after-image
    representation.

The responsibility of diffing the two snapshots is still up to the
SnapshotComparer, but rendering the pixels is a matter dealt with by the
subclasses.

Instead of creating one diff image to rule them all, we now create three
images. The idea is that you will be able to quickly toggle between the
different representations on the snapshot#show page. We have found this
to be a better UI for quickly reviewing all the different diffs that can
arise.

- The overlay is great for quickly spotting where changes have been
  made.
- By toggling between the before- and after-image, you can easily tell
  what changed.

I choose to remove the gutter for now. I'm not sure that's the best
idea, but I'd like to get some testing done first. If we choose to
reintroduce the gutter, we can render it through a subclass of
SnapshotComparisonImage.