Fix RED and GREEN color constant values
When I first converted these hex colors to the values that ChunkyPNG
likes, I must have crossed a wire and put the values in the wrong spots.
This made the RED show up as green, and the GREEN show up as red, which
was confusing when the images were actually being rendered. I had
originally "fixed" this by reversing the before and after images as they
were sent to the differ, but this was ultimately incorrect.

Change-Id: Ie752772aacc6b198a4178db3d2aca700c5d531a0