Add basic i18n
For now I added a check for a query string parameter to set the locale.
To change the locale, just use the `locale` param.

Since we don't have any concept of an account or a person's preferences,
we may want to check the Accept-Language HTTP header to determine which
language to use. More info on using the Accept-Language HTTP header can
be found at:

  http://guides.rubyonrails.org/i18n.html#setting-the-locale-from-the-client-supplied-information

For now, this is just a quick first pass at extracting copy into an
English configuration file. Next steps include extracting more copy and
adding additional translations.

Addresses #35

Change-Id: I314331dd373dfb6309f50633dd98320b040b8b96