Move and condense snapshot page title and url
The content here should speak for itself, no need to put it in a table
with clear labels. Moving it up and putting it on one line helps to
remove clutter and condense the preamble of each snapshot page, allowing
more of the diff image to be seen above the "fold".

This happens to remove the keyboard focusable property of the URL, but I
don't think that was a valuable enough element to be keyboard focusable
so I'm not too worried about losing it.

Change-Id: Ie5156e31d865e6e58fa02544cb372887d340b6f5