Improve breadcrumb for snapshots not part of sweeps
When in a sweep, you usually want to hit "u" to go from the snapshot to
the sweep, not to the URL page. However, if you are looking at a
snapshot that is not part of a sweep, you want to go to the URL page,
not the project page. This commit makes that happen.

At the same time, I made this link use the URL's page title, which is
consistent with the table on the project page. I wasn't sure if we
should do the same for the breadcrumb when part of a sweep, so I'll
leave that decision up for another day.

Change-Id: I3724ad83c7728cd23cf0012c38315cd3c36902ed