Fix bug with rendering diff when after is wider than before
There was a bug in our diff image rendering that prevented the
additional pixels from being rendered on changed rows when the after
image is wider than the before image. This was happening because
Array#zip maintains the length of the original array, causing the extra
pixels to be thrown away. To fix this, we just need to detect that this
is happening and render the pixels as added.

Change-Id: I47742d6bdf7e7b1238d8567ba05b04d1b1b4e9cf