Move snapshot page header into partial
In 6698b9b2da677bd4 I removed the status row from the snapshot page in
an effort to make it shorter. However, as @trotzig pointed out, the
pending state is now difficult to identify.

To resolve this, I am going to bring back a status indicator, but this
time it will be part of the page header and done in a more compact way
than we had done before.

This commit brings us one step closer to this vision by moving the page
header into the partial that is rendered when the accept/reject buttons
are activated. As of this commit, this has no real purpose, but my next
commit will add a status indicator here that will need to be refreshed
when the status changes.

Change-Id: I3e6295497bd8deaab181e51e3a2d4194ca79b0f4