Move re-take and re-compare buttons to header
These buttons are rarely used and more similar in nature to the delete
button, so it makes more sense to represent with just icons and put them
all near each other. At the same time, I changed the icons for these two
buttons, since they were previously the same, and I made the delete
button an actual button that looks like a link instead of a link that
acts like a button. I also added title text to help people understand
the effects of these buttons in case the icons aren't clear enough.

Change-Id: Ic321306cc77be0c09dd8dc00129ad1dc08bb949e