Move snapshot buttons into header
If the snapshot is under review, accepted, or rejected, we don't need to
double up the indicator and the buttons because they show the same
information. Instead, we can just put the buttons where the status
indicator goes. This also ends up saving vertical space, which will
nicely push the diff higher above the "fold".

At the same time, I noticed a typo in one of the status indicator class
names, so I fixed it.

Change-Id: I11b93669ad482ffb51ff0deda5a926b5c4617e6f