Use sdiff size instead of max height for output images
The number of rows in the resulting diff will, in many cases, be larger
than the number of rows in both the before and after images. Therefore,
we need to use this size to determine the height of the output images.
If we do not do this, the output images will not show the content all
the way to the bottom if there are any "added" or "deleted" rows.

We may want to make the before and after comparison images to use the
widths of the before and after images, in case their widths shift, but
depending on how we end up rendering these, that may not actually be
desirable.

At the same time, I changed the diff percentage calculation to use this
as well, helping to keep the diff percentage under 100%.

Change-Id: I3482016ad776a00bad43cf5408b52249a02c55d5