Add tab navigation for snapshot diff sprite
Now that we save the snapshot diff image as a sprite consisting of three
images, we need to make the rendering a bit smarter. By adding tabs, and
only showing a third of the image at once, we allow a more focused
review. Most of the time, the "Diff" image is enough to make a good
decision about whether or not the diff should be accepted. This is also
the part of the image that we default to.