Fix diff percentage
Now that we've delegated diff image creation to
`SnapshotComparisonImage` and its subclasses, the `diff_in_percent`
value broke. We were computing the percentage by dividing the number of
lines changed with the total number of pixels in the image.

This commit moves the responsibility of computing this value to the
`DiffClusterFinder`, and changes it to be number of different lines
divided by the total number of lines.