Base the diff percentage score on `sdiff.size`
We were calculating the percentage diff score for a snapshot based on
number of changed lines divided by the max height of the before and
after image. This had the potential of making the diff larger than 100%
since the number of changed lines can actually exceed the height of any
of the images. This would be the case if there was a lot of added and
deleted lines. We are now instead using the total count of lines that
`Diff::LCS.sdiff` reports.

This had the added benefit of removing the need for the max_height
altogether.