Render additions/deletions without color overlay
Now that we have the gutter in a better state, we don't need the
green/red overlay on the before/after image to display
additions/deletions.

This has been ping-ponged a bit back and forth, and it's possible that
we'll bring it back. But for now, this is the our best assumption about
how to make the experience as good as possible.