Show Photoshop-style checkered pattern where there are gaps
To better visualize that there is a gap in the before or after image, we
render a checkered, Photoshop-style, background pattern. The gaps were
already there, all I had to do was to make sure that the overlayed image
was faded on a white background instead of a transparent, and that the
gutter didn't render anything transparent.