Show gray rulers for each diff cluster, when focused
To aid looking at a diff cluster at a time, I'm adding a gray border on
top and bottom of the cluster when it is focused.

This haml code is starting to get ugly, and we should probably start
simplifying it soon. But for now I think this is fine.