Vertically center the diff cluster arrow
Before, the arrow would be rendered close to the top of the cluster.
That was ok, at least for clusters that were taller than the icon. But
when the cluster was smaller than the icon, it felt weird. We're now
centering it vertically.

The only place where this might be a problem is when the cluster is
taller than twice the size of the screen viewport. The icon would then
render outside the visible area. But those clusters are complicated
enough already, so having the arrow in the wrong place just makes it
marginally worse.