Set animation-duration to 0 when taking snapshots
When we added some styles and JS to disable jQuery animations and CSS
transitions, we had forgotten about CSS animations. This commit adds
some styles to the page as it is snapshotted to attempt to disable CSS
animations by setting their duration to 0. This will allow the
animations to complete, but they will happen instantly. This will help
ensure that screenshots are more consistent on sites that use this
feature, reducing the noise.

Change-Id: Iaf942231e0316872dfdd74d4adba03efa5170256