Memoize more things in SnapshotComparisonImage::Overlayed
In 0b3aba2c4366a we added some memoization to the overlayed image
generation, which gave us some pretty good performance improvements. As
noted there, I thought we could get even more performance benefits by
moving this memoization up the chain. This commit does that, preventing
repeated calls to `fade` in addition to `compose_quick`.

In my extremely non-scientific test, this seems to give us roughly a 25%
improvement. Of course, this improvement is highly dependent on the
specifics of the before and after images.

Addresses #76.

Change-Id: I714d6c4eaa05b3d53f77cdbc40cf21a80648c90f