Improve scrolling performance on snapshot pages
After affixing the snapshot pills, we introduced some scrolling
animation jankiness. This was happening because the browser needed to
repaint a bunch of stuff while scrolling. In Chrome, we can use a null
translate hack to force a layer so the browser will use the much faster
layer compositing and avoid the expensive painting.

Change-Id: Ie5a67e4d51f388d4a101406bf603f1a480f18489