Flip faded pixel strategy
Previously we were fading the alpha channel of the unchanged pixel in
the overlay image, and then composing it on top of a white color.
Instead, we can eek out a little better performance by flipping this.
Now we fade the white color once and compose it on top of the fully
opaque pixel color. The end result is the same (except for pixels that
are not actually opaque), but a little faster.

Closes #76.

Change-Id: I5da782156804d417f1ce2279812318913c8dcbdd