Use white gradient to make diff pills visible
In 49ebb1ae28867a5472fd we added a box shadow to help make these
elements more readable on various backgrounds. This commit takes that
approach to the next level by using a white to transparent gradient.
This gives the top of the page a faded out look, which makes the
selected and non-selected pills nicely readable.

Change-Id: I02ebc3e156af9299e99e6f899a744db5e92bb8b9