Make diff percentage specs more specific
Now that we have our diff percentages settled out a little better than
before, it feels like the right time to make these specs more specific.
This will give us better insight as to when these values may change as
we modify the code.

Worth noting is the somewhat unintuitive fact that the diff percentages
for images that are simply wider or narrower are 100% diffs. This is
because we are currently calculating the diff percentages based on the
number of rows that have changed. And, when the image changes width,
every row has changed. We may decide to change this behavior at some
point, but this is how it currently works.

Change-Id: I5405127cdb582e681da77a43ff3cbcd764746e9a