Use <meter> for diff percentages in snapshot thumbnails
I felt like these diff percentages weren't very useful when viewing a
bunch of snapshot thumbnails. I never looked at the numbers, so I wasn't
getting very much benefit from their existence. I'd like to make this
information more consumable, so I am replacing the text with a more
visual representation. I decided to use the <meter> element and styled
it to match the rest of the app.

I chose the values of low and high fairly arbitrarily, so we'll see how
things shake out as we get accustomed to seeing this information
presented in this way.

I have the concern that people will not understand what this meter
represents, but I think we could also add it to the snapshot page itself
where it could be presented next to the actual numerical value, to help
build a connection.

I am also worried that we now have too many bar-like elements floating
around, and that may get confusing as to what they actually mean. I
think we can probably resolve this with the correct styling over time,
so I am not going to worry about it too much.

Change-Id: I684ab42e4ce96e897ac4f1b8e36e851f8fd0f1f8