Add friendly intro jumbotron to main page
When people first set up Diffux, they were previously presented with a
more or less blank page. While this worked, it wasn't very friendly. I'd
like to start people off on the right foot, so if they don't have any
projects yet, I think it would be nice to show a friendly message
welcoming them to the app and gives them a link to help encourage
contributions.

The copy that I wrote is just a pretty rough first draft, but I figured
it was better than what we had before, so I'm shipping it.

Change-Id: I97e5241497d82dc435591a5cd9a7a7923d8b7988