Remove redundant `#update_counters!`
When snapshots for a sweep change, we call `Sweep#refresh!` to update
counters and perhaps send out an email if the sweep is done. In the code
that sends out the email, there was a redundant call to update the
counters again. This might be the reason to bug #83. I've removed the
redundant call and updated specs to be more like the live code - instead
of calling private methods directly, we call `#refresh!`.

There are a couple of redundant saves here too that I'm investigating
next.