Lock `Sweep#refresh!` to one thread at a time
A couple of bugs are indicating that there's a lot of race issues in
this method. My recent two commits have probably made it less likely
for these to happen but I wanted to follow up with something that has a
better potential of getting rid of the problem altogether.

I'm using a pessimistic block (`with_block`) to lock multiple threads
from calling `#refresh!` at once. I was first going to use a local
mutex, but since it is very likely that workers are run on more than one
machine, this wasn't going to scale. A database level lock should be
more robust and is exactly what `with_lock` gives us.

More on pessimistic locking:
http://api.rubyonrails.org/classes/ActiveRecord/Locking/Pessimistic.html