Make pending snapshots part of the sweep progress bar
Before, the progress bar wouldn't start showing anything interesting
until the first snapshot image had been taken. After this commit, it
will immediately go entirely gray and become active. I made this happen
by putting the pending count last in the list of stacked progress bars,
and changed the background color to gray. It now looks like the progress
bar is in an indeterminate state to begin with, which is better in my
opinion.