Always show the snapshot image on the snapshot#show page
A recent commit (8b94b558094) removed the before and after image from
the snapshot#show page. This was a win in case the snapshot was in a
state where it had a non-zero diff because we saved a lot of space by
not rendering the before and after images. However, when the snapshot
had a zero diff and when it didn't have a baseline to compare against,
you wouldn't see the snapshot image at all.

I added the image back in two places. I also added an alert to explain
the state when there is no baseline to compare against.

The spec coverage for all of this is not great. But I did add some smoke
tests just to make sure that the page would render with snapshots in
different states. And while doing that I found a bug in the snapshot
factory that I fixed.