Fix off-by-one error in snapshot_diff_cluster.rb
We were off by one pixel when calculating percentages in this file. As
pointed out by @lencioni in 1474692316688b3913.

I had to wrap my head around this issue once or twice (pun intended).
The off-by-one was obvious, but constructing the specs in a way that
made sense was harder than I thought.

This closes #80.