Fix failing snapshots_controller_spec
I added a help modal for keyboard shortcuts that made specs fail for the
snapshots controller. The specs weren't specific enough so some of them
picked up copy from the help modal and started failing. I fixed the bug
by being more precise in the specs (looking for elements with a certain
class instead of copy on the page).

To make testing easier by having access to a `have_css` matcher, I added
capybara to the list of test dependencies.