Don't send email when email address is empty
We noticed that the sweep progress bar stopped updating after a while.
By running `sweep.refresh!` manually we noticed that the code was
failing due to the email address being an empty string. This commit
mitigates that by changing the condition in where we return from the
`send_email_if_needed!` method in case the email is missing.

This fixes issue #87.