Draw only one ChunkyPNG::Image in SnapshotComparer
As pointed out by @lencioni in issue #76, we are spending a lot of time
stitching together the individual images created by the different
`SnapshotComparisonImage`s. This commit changes the implementation of
these classes to reuse the same image canvas for all comparison images,
but use an x-offset to control where on the x-axis the individual images
will start to draw.