Constantize some math from SnapshotComparisonImage#pixel_diff_score
In #pixel_diff_score, we take the Euclidean distance between the RGBA
channels of the 2 given colors over the maximum possible Euclidean
distance. This gives us a percentage of how different the two colors
are.

Since the maximum possible Euclidean distance never changes, we can
avoid performing the math (an exponent operation, a multiplication
operation, and a square root operation) each time we want to get the
pixel diff score of two colors. Since this has the potential of
happening many many times, this should be a nice, little performance
win.

Change-Id: I90fd884c231e1a87af2b7697ab223293ca851918