Un-focus any focused elements before taking snapshots
We've been taking snapshots of http://sassmeister.com/ and have noticed
that we're randomly getting diff on the cursor of the focused text input
field. The blinking cursor is creating diffs randomly, depending on when
during the blink cycle the snapshot is taken.

We believe that we can avoid this by un-focusing any focused elements
before taking the snapshot.

Closes #92.

Change-Id: I41b122fb1c073dc6dbc342954792d9177d054555