Make `viewport_widths` and `url_addresses` fields required
This doesn't fully resolve issue #90 and #91, but it takes care of some
of it.  Before this change, you could create a project without any URL
addresses and/or viewport widths. This is still the case, even after
this change (we parse these fields on the server and ignore anything
that's unparseable). Making the fields required fixes the problem where
you could save a project with these fields blank.