Add "Last sweep" box to start page
It is convenient to have a direct link to the last sweep that was
made/is running. I made the projects#index page two-column and put a box
with the last sweep displayed.

While in the test I improved a couple of `should include` validations to
start using capybara matchers instead. I think these matchers are more
closely matched to the intention of the spec.