Stop all jQuery animations before taking snapshots
Setting `$.fx.off` prevents future animations from happening, but has
no effect on animations that are happening when it is set. To make sure
that all of these animations are in their final resting places, I added
a bit of jQuery that stops all animations on all elements, clears their
animation queues, and jumps to the ends of all of these animations. This
will help our snapshots be more consistent.

While I was at it, I adjusted the guards on `$` and `$.fx` to simply
check for the existence of jQuery. This should be good enough. I also
switched our usage of `$` to `jQuery`, since `$` may be defined as
something other than jQuery on some sites, while `jQuery` is likely only
going to be jQuery.

Change-Id: I77122c0db8be5f0985460e4717201d055a6410c8